Started moving some unit tests over to e2e tests.
I'm going to have to add a couple of filters to the view for the paging, and while the filters themselves are unit tested, mocking each of them out for the view rendering unit tests was starting to get messy, and probably a touch difficult to maintain.

An e2e smoke test on this functionality will ensure the linkage is there, and the end user sees what is expected.  e2e test at this stage is just a skeleton.